removed api function mysql_get_server_status (introduced in last commit). Instead of mariadb_get_infov now supports additional parameters: * MARIADB_CONNECTION_SERVER_STATUS * MARIADB_CONNECTION_SERVER_CAPABILITIES * MARIADB_CONNECTION_EXTENDED_SERVER_CAPABILITIES * MARIADB_CONNECTION_CLIENT_CAPABILITIES

+++ plugins/auth/dialog.c
@@ -138,7 +138,7 @@ static int auth_dialog_open(MYSQL_PLUGIN_VIO *vio, MYSQL *mysql)
   my_bool first_loop= TRUE;
 
   do {
-    if ((packet_length= vio->read_packet(vio, &packet)) == (size_t)-1)
+    if ((packet_length= vio->read_packet(vio, &packet)) == -1)
       /* read error */
       return CR_ERROR;
